refactor: separate types into internal and external
